1. Eliminate legacy chunked as-of join state managers and clean up code accordingly. 2. Fix position overflow bug in left hash slots/redirections gathering during as-of join instantiation. 3. Switch to a narrower type for hash slot gathering in AJ since we no longer need space for overflow masking. 4. Finish deleting the IncrementalChunkedNaturalJoinStateManager and making a few more long->int conversions for NJ.

+++ engine/table/src/main/java/io/deephaven/engine/table/impl/NaturalJoinHelper.java
@@ -718,7 +718,7 @@ class NaturalJoinHelper {
         }
 
         @Override
-        public void accept(long updatedSlot, long originalRightValue, byte flag) {
+        public void accept(int updatedSlot, long originalRightValue, byte flag) {
             final RowSet leftIndices = jsm.getLeftIndex(updatedSlot);
             if (leftIndices == null || leftIndices.isEmpty()) {
                 return;
